tests/misc/update_loses_autoinstalled_flag.py: update test
- Remove old bug switch
- Only execute last case when using the internalsolver. Upgrade
behavior on libsolv doesn't remove packages, which makes the case not
applicable to libsolv.

+++ tests/misc/update_loses_autoinstalled_flag.py
@@ -3,9 +3,10 @@
 import os
 import opk, cfg, opkgcl
 
-bug = True
-
 opk.regress_init()
+
+_, version = opkgcl.opkgcl("--version")
+libsolv = True if 'libsolv' in version else False
 
 open("asdf", "w").close()
 a = opk.Opk(Package="a", Version="1.0", Depends="b")
@@ -24,36 +25,38 @@ opkgcl.install("a")
 if not opkgcl.is_autoinstalled("b"):
 	opk.fail("b is not autoinstalled")
 
-if (bug):
-	a = opk.Opk(Package="a", Version="2.0", Depends="b")
-	a.write()
-	b = opk.Opk(Package="b", Version="2.0")
-	b.write(data_files=["asdf"])
-
-	o = opk.OpkGroup()
-	o.addOpk(a)
-	o.addOpk(b)
-	o.write_list()
-
-	opkgcl.update()
-	opkgcl.upgrade();
-
-	if not opkgcl.is_autoinstalled("b"):
-		opk.fail("b is not autoinstalled anymore")
-
-a = opk.Opk(Package="a", Version="3.0")
-a.write(data_files=["asdf"])
-os.unlink("asdf")
+a = opk.Opk(Package="a", Version="2.0", Depends="b")
+a.write()
+b = opk.Opk(Package="b", Version="2.0")
+b.write(data_files=["asdf"])
 
 o = opk.OpkGroup()
 o.addOpk(a)
+o.addOpk(b)
 o.write_list()
 
 opkgcl.update()
 opkgcl.upgrade();
 
-if opkgcl.is_installed("b", "2.0"):
-	opk.xfail("[libsolv] b is still installed")
+if not opkgcl.is_autoinstalled("b"):
+	opk.fail("b is not autoinstalled anymore")
 
-if not opkgcl.is_installed("a", "3.0"):
-	opk.fail("a is not installed")
+# The internalsolver is more agressive removing orphans than libsolv. On libsolv upgrades, no
+# packages are removed (on dist-upgrade, removal is allowed)
+if not libsolv:
+    a = opk.Opk(Package="a", Version="3.0")
+    a.write(data_files=["asdf"])
+    os.unlink("asdf")
+
+    o = opk.OpkGroup()
+    o.addOpk(a)
+    o.write_list()
+
+    opkgcl.update()
+    opkgcl.upgrade();
+
+    if opkgcl.is_installed("b", "2.0"):
+        opk.fail("b is still installed")
+
+    if not opkgcl.is_installed("a", "3.0"):
+        opk.fail("a is not installed")
